Fix for server_password management
The fix provided by #53 is incomplete: using shell redirections and
control operators need to switch to the 'shell' exec provider, which in
turn allows command injections.  Moreover, as explained in #52, murmurd
exists with an exit code of 1 on success…  but also on failure.

Attempt to improve the situation by switching to the shell provider,
shell escaping the password thanks to stdlib's shell_escape function,
and grep the stderr output of murmurd for a success message.